# SWAGGER/BankingAPISwagger.yaml
openapi: 3.0.0
info:
  title: Banking API
  description: Banking API Group 1 Endpoints
  version: "1.0.0"
  contact:
    email: 683029@student.inhollan.nl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/JOANIDEVELOP/BankingAPIProject/1.0.0
paths:
...
  /accounts/own/transfers:
    post:
      tags:
        - Customer
      summary: Transfer funds between user's own accounts
      operationId: transferBetweenOwnAccounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        '201':
          description: Transfer successfully completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'

        '400':
          description: Invalid transfer details, incomplete or invalid data.
        '404':
          description: Source or destination account not found.
        '422':
          description: Transfer cannot be completed, insufficient funds or account restrictions.
      security:
        - bearerAuth: []
  /accounts/{accountId}:
    delete:
      tags:
        - Employee
      summary: Close account
      description: This can only be done by the employee.
      operationId: closeAccount
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
            format: int64
            example: 123
            description: ID of the account to be closed
      responses:
        '200':
          description: Account closed successfully
          content:
            application/json:
              schema:
                type: string
                example: "Account was closed successfully"
        '404':
          description: Account not found
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        '500':
          description: Internal server error
      security:
        - bearerAuth: [ ]
...
  /accounts/{userId}/details:
    get:
      tags:
        - Customer
      summary: Get account and user details for a user
      operationId: getAccountAndUserDetails
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User and account details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CombinedUserAccount'
        '404':
          description: User ID not found.
      security:
        - bearerAuth: []
...
  /accounts/customers:
    get:
      tags:
        - Employee
      summary: Get all customer accounts
      description: Get all customer accounts
      operationId: getAllCustomerAccounts
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"

        "400":
          description: Invalid status value.
        '401':
          description: User is not authorized
        '403':
          description: User must be of role employee
    post:
      tags:
        - Employee
      summary: Create account
      description: Create an account for a customer
      operationId: createAccount
      requestBody:
        description: Created account object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
            examples:
              Account789e4567-e89b-12d3-a456-426614174000:
                $ref: "#/components/examples/Account789e4567-e89b-12d3-a456-426614174000"
        required: true
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid status value.
        '401':
          description: User is not authorized
        '403':
          description: User must be of role employee
      security:
        - bearerAuth: []
    put:
      tags:
        - Employee
      summary: Update account
      description: This can only be done by the employee
      operationId: updateAccount
      requestBody:
        description: Updated account object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
            examples:
              Account789e4567-e89b-12d3-a456-426614174000:
                $ref: "#/components/examples/Account789e4567-e89b-12d3-a456-426614174000"
        required: true
      responses:
        "400":
          description: Invalid account supplied
        "404":
          description: Account not found
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
      security:
        - bearerAuth: []
  /accounts/savings/{userId}:
    get:
      tags:
        - Employee
      summary: Get all savings accounts of the user
      description: This can only be done by the employee.
      operationId: getSavingsAccountsOfUser
      parameters:
        - name: userId
          in: path
          description: ID of user which owns the accounts
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

        "400":
          description: Invalid status value.
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "404":
          description: User not found
      security:
        - bearerAuth: []
  /accounts/checking/{userId}:
    get:
      tags:
        - Employee
      summary: Get all checking accounts of the user
      description: This can only be done by the employee.
      operationId: getCheckingAccountsOfUser
      parameters:
        - name: userId
          in: path
          description: ID of user which owns the accounts
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

        "400":
          description: Invalid status value.
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "404":
          description: User not found
      security:
        - bearerAuth: []
  /accounts/byAbsoluteLimit:
    get:
      tags:
        - Employee
      summary: Get all accounts with an absolute limit less than or equal to a specific limit
      description: This can only be done by the employee.
      operationId: getAccountsByAbsoluteLimit
      parameters:
        - name: absoluteLimit
          in: query
          description: the absolute limit applicable for an account
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

        "400":
          description: Invalid status value.
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
      security:
        - bearerAuth: []
  /accounts/inactive:
    get:
      tags:
        - Employee
      summary: Get all accounts which were set as inactive by the bank
      description: This can only be done by the employee.
      operationId: getInactiveAccounts
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  examples:
...
    Account789e4567-e89b-12d3-a456-426614174000:
      value:
        id: 789e4567-e89b-12d3-a456-426614174000
        userId: 789e4567-e89b-12d3-a456-426614174000
        IBAN: NL91ABNA0417164300
        currency: EUR
        accountType: savings
        isActive: true
        balance: 10250.50
        absoluteLimit: 0
  schemas:
    CombinedUserAccount:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        account:
          $ref: '#/components/schemas/Account'
    User:
      type: object
      required:
        - id
        - username
        - email
        - firstName
        - lastName
        - BSN
        - phoneNumber
        - birthDate
        - totalBalance
        - role
        - dailyLimit
        - isApproved
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "johndoe78"
        email:
          type: string
          format: email
          example: "johndoe@example.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        BSN:
          type: string
          example: "999999982"
        phoneNumber:
          type: string
          example: "+31612345678"
        birthDate:
          type: string
          format: date
          example: "1985-04-12"
        totalBalance:
          type: number
          format: double
          example: 3500.75
        role:
          type: string
          enum:
            - customer
            - employee
          example: "customer"
        dailyLimit:
          type: number
          format: double
          example: 500.00
        isApproved:
          type: boolean
          example: true
    Transaction:
      type: object
      required:
        - id
        - fromAccount
        - toAccount
        - amount
        - date
        - userId
      properties:
        id:
          type: string
          format: uuid
          example: "223e4567-e89b-12d3-a456-426614174000"
        fromAccount:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        toAccount:
          type: string
          format: uuid
          example: "321e4567-e89b-12d3-a456-426614174000"
        amount:
          type: number
          format: double
          example: 250.00
        date:
          type: string
          format: date-time
          example: "2023-10-05T14:48:00.000Z"
        userId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
    Account:
      type: object
      required:
        - id
        - userId
        - IBAN
        - currency
        - accountType
        - isActive
        - balance
        - absoluteLimit
      properties:
        id:
          type: string
          format: uuid
          example: "789e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        IBAN:
          type: string
          example: "NL91ABNA0417164300"
        currency:
          type: string
          example: "EUR"
        accountType:
          type: string
          example: "savings"
        isActive:
          type: boolean
          example: true
        balance:
          type: number
          format: double
          example: 10250.50
        absoluteLimit:
          type: number
          format: double
          example: 0